test: rewrite RemoteError tests
