--- test/fixtures/remote-error-test-cases/default-messages.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const jstp = require('../../../');
-
-module.exports = Object.keys(jstp)
-  .filter(key => key.startsWith('ERR_'))
-  .map(key => jstp[key].toString());
